pch_i2c: bail out if BAR isn't configured
Reviewed-on: https://review.haiku-os.org/c/haiku/+/4883
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ src/add-ons/kernel/busses/i2c/pch_i2c.h
@@ -72,7 +72,7 @@ typedef enum {
 
 typedef struct {
 	phys_addr_t base_addr;
-	size_t map_size;
+	uint64 map_size;
 	uint8 irq;
 	i2c_bus sim;
